// apps/wear/pushtracker/app/enums/notification-keys.ts
export enum NotificationKeys {
  DAILY_PUSH_WARNING = 'DAILY_PUSH_WARNING_NOTIFICATION',
  NEW_COAST_RECORD = 'NEW_COAST_TIME_RECORD_NOTIFICATION',
  ADD_PT_ELEMENT = 'ADD_PUSHTRACKER_ELEMENT_NOTIFICATION',
  WC_MAINTENANCE = 'GENERAL_WC_MAINTENANCE_NOTIFICATION'
}

/**
 * IDs for PushTracker Wear Notifications
 */
export enum PushTrackerNotificationIDs {
  DAILY_PUSH_WARNING = 111111,
  NEW_COAST_TIME_RECORD = 111112,
  ADD_PT_AS_ELEMENT = 111113,
  GENERAL_WC_MAINTENANCE_REMINDER = 111114
}